export registerServerComponent as a separate entrypoint to avoid client bundle increase
