+++ knip.ts
@@ -9,6 +9,7 @@ const config: KnipConfig = {
         'node_package/src/ReactOnRails.node.ts!',
         'node_package/src/ReactOnRailsRSC.ts!',
         'node_package/src/RSCWebpackLoader.js!',
+        'node_package/src/registerServerComponent.js!',
       ],
       project: ['node_package/src/**/*.[jt]s!', 'node_package/tests/**/*.[jt]s'],
       babel: {
