+++ node_package/src/ReactOnRails.ts
@@ -19,7 +19,6 @@ import type {
   AuthenticityHeaders,
   Store,
   StoreGenerator,
-  RegisterServerComponentOptions,
 } from './types';
 import reactHydrateOrRender from './reactHydrateOrRender';
 
@@ -53,16 +52,6 @@ ctx.ReactOnRails = {
    */
   register(components: { [id: string]: ReactComponentOrRenderFunction }): void {
     ComponentRegistry.register(components);
-  },
-
-  /**
-   * Register a specific component as a server component.
-   * The component will not be included in the client bundle.
-   * When it's rendered, a call will be made to the server to render it.
-   * @param componentNames
-   */
-  registerServerComponent(options: RegisterServerComponentOptions, ...componentNames: string[]): void {
-    ComponentRegistry.registerServerComponent(options, ...componentNames);
   },
 
   registerStore(stores: { [id: string]: StoreGenerator }): void {
