+++ node_package/src/types/index.ts
@@ -160,7 +160,6 @@ export type RenderReturnType = void | Element | Component | Root;
 
 export interface ReactOnRails {
   register(components: { [id: string]: ReactComponentOrRenderFunction }): void;
-  registerServerComponent(options: RegisterServerComponentOptions, ...componentNames: string[]): void;
   /** @deprecated Use registerStoreGenerators instead */
   registerStore(stores: { [id: string]: StoreGenerator }): void;
   registerStoreGenerators(storeGenerators: { [id: string]: StoreGenerator }): void;
