+++ spec/dummy/config/webpack/clientWebpackConfig.js
@@ -9,6 +9,10 @@ const configureClient = () => {
   // client config is going to try to load chunks.
   delete clientConfig.entry['server-bundle'];
 
+  // clientConfig.resolve.fallback = {
+  //   ...clientConfig.resolve.fallback,
+  //   'process/browser': false,
+  // };
   return clientConfig;
 };
